refactor(kotest): hoist toRawRequest into wiremock integration, drop example transport
Move the WireMock Request -> Wirespec.RawRequest mapping out of the example's
ProjectConfig into the shared wiremock integration module as a public
Request.toRawRequest(), and point the example at the integration-provided
HttpTransportation instead of its own copy.
